--- a/src/vm.c
+++ b/src/vm.c
@@ -773,11 +773,17 @@ DEFINE_PRIMITIVE("%call-for-values", call_for_values, subr1, (SCM prod))
     case 0: return STk_nil;
     case 1: return LIST1(vm->val);
     default:  {
-                SCM  res = STk_nil;
                 if (len <= MAX_VALS) {
-                  for (int i = len-1; i >= 1; i--)
-                    res = STk_cons(vm->vals[i], res);
-                  return STk_cons(vm->val, res);
+                  /* vm->val goes into CAR(res), and
+                     vm->vals go into the other list positions: */
+                  SCM res = STk_C_make_list(len, STk_nil);
+                  SCM ptr = CDR(res);
+                  CAR(res) = vm->val;
+                  for (int i = 1; i <= len-1; i++) {
+                    CAR(ptr) = vm->vals[i];
+                    ptr = CDR(ptr);
+                  }
+                  return res;
                 } else {
                   return STk_vector2list(vm->vals[0]);
                 }