This makes processing of variable args faster.

```scheme
(import (scheme list))
(define (f . rest) 1)
(define-macro (call-f-with-500-args)
  (append '(f) (iota 500)))

(time (repeat 100000 (call-f-with-500-args)))
```

Old code:
2130.932 ms
50000000 allocation

New code:
1660.077 ms
100000 allocation

```scheme
(define-macro (call-f-with-2-args)
  (append '(f) (iota 2)))

(time (repeat 10000000 (call-f-with-2-args)))
```

Old code:
1256.636 ms
20000000 allocation

New code:
1159.518 ms
10000000 allocation