@jpellegrini 2025-04-14
> For example, I hadn't time to work on it, but it seems that `%call-for-values` also conses.

I added a commit for this...